+++ src/dakara_player/font_loader.py
@@ -1,10 +1,9 @@
 """Load fonts on user level for the media players."""
 
+import ctypes
 import logging
-import os
 import sys
 from abc import ABC, abstractmethod
-from contextlib import ExitStack
 
 from path import Path
 
@@ -14,6 +13,7 @@ try:
 except ImportError:
     from importlib_resources import contents, path
 
+from dakara_base.exceptions import DakaraError
 
 logger = logging.getLogger(__name__)
 
@@ -41,11 +41,19 @@ class FontLoader(ABC):
     """Abstract font loader.
 
     Must be specialized for a given OS.
+
+    Args:
+        package (str): Package checked for font files.
+
+    Attributes:
+        package (str): Package checked for font files.
     """
 
-    GREETINGS = "Dummy font loader selected"
+    GREETINGS = "Abstract font loader selected"
 
-    def __init__(self):
+    def __init__(self, package):
+        self.package = package
+
         # show type of font loader
         logger.debug(self.GREETINGS)
 
@@ -60,28 +68,44 @@ class FontLoader(ABC):
     def __enter__(self):
         return self
 
-    def __exit__(self, exception_type, exception_value, traceback):
+    def __exit__(self, *args, **kwargs):
         self.unload()
 
     def get_font_name_list(self):
-        """Extract font names in font directory.
+        """Give name of package fonts.
 
         Returns:
             list of str: List of font names.
         """
         logger.debug("Scanning fonts directory")
-        return [
+        font_file_name_list = [
             file
-            for file in contents("dakara_player.resources.fonts")
+            for file in contents(self.package)
             if Path(file).ext.lower() in FONT_EXTENSIONS
         ]
+        logger.debug("Found %i font(s) to load", len(font_file_name_list))
+
+        return font_file_name_list
+
+    def get_font_path_iterator(self):
+        """Give font paths in font package.
+
+        Yields:
+            path.Path: Absolute path to the font, from the package.
+        """
+        for font_file_name in self.get_font_name_list():
+            with path(self.package, font_file_name) as font_file_path:
+                yield Path(font_file_path)
 
 
 class FontLoaderLinux(FontLoader):
     """Font loader for Linux.
 
-    It symlinks fonts to load in the user fonts directory. On exit, it
-    removes the created symlinks.
+    It copies fonts to load in the user fonts directory and removes them on
+    exit. Using symbolic links is not safe as the location of the package fonts
+    may not be permanent (see `importlib.resources.path` for more info).
+
+    See: https://docs.python.org/3/library/importlib.html#importlib.resources.path
 
     Example of use:
 
@@ -89,55 +113,154 @@ class FontLoaderLinux(FontLoader):
     ...     loader.load()
     ...     # do stuff while fonts are loaded
     >>> # now fonts are unloaded
+
+    Args:
+        package (str): Package checked for font files.
+
+    Attributes:
+        package (str): Package checked for font files.
+        font_loaded (dict of path.Path): List of loaded fonts. The key is the
+            font file name and the value is the path of the installed font in
+            user directory.
     """
 
     GREETINGS = "Font loader for Linux selected"
-    FONT_DIR_SYSTEM = Path("/usr/share/fonts")
-    FONT_DIR_USER = Path("~/.fonts")
+    FONT_DIR_SYSTEM = Path("/") / "usr" / "share" / "fonts"
+    FONT_DIR_USER = Path("~") / ".fonts"
 
-    def __init__(self):
+    def __init__(self, *args, **kwargs):
         # call parent constructor
-        super().__init__()
+        super().__init__(*args, **kwargs)
 
         # create list of fonts
-        self.fonts_loaded = []
+        self.fonts_loaded = {}
 
     def load(self):
         """Load the fonts."""
         # ensure that the user font directory exists
-        try:
-            os.mkdir(self.FONT_DIR_USER.expanduser())
+        self.FONT_DIR_USER.expanduser().mkdir_p()
 
-        except OSError:
-            pass
+        # get system and user font files
+        system_font_path_list = list(self.FONT_DIR_SYSTEM.walkfiles())
+        user_font_path_list = list(self.FONT_DIR_USER.expanduser().walkfiles())
 
         # load fonts
-        self.load_from_resources_directory()
+        for font_file_path in self.get_font_path_iterator():
+            self.load_font(font_file_path, system_font_path_list, user_font_path_list)
 
-    def load_from_resources_directory(self):
-        """Load all the fonts situated in the resources font directory."""
-        font_file_name_list = self.get_font_name_list()
-
-        logger.debug("Found %i font(s) to load", len(font_file_name_list))
-        self.load_from_list(font_file_name_list)
-
-    def load_from_list(self, font_file_name_list):
-        """Load the provided list of fonts.
+    def load_font(self, font_file_path, system_font_path_list, user_font_path_list):
+        """Load the provided font.
 
         Args:
-            font_file_name_list (list of str): List of name of the fonts to
-                load.
+            font_file_path (path.Path): Absolute path of the font to load.
+            system_font_path_list (list of path.Path): List of absolute paths
+                of system fonts.
+            user_font_path_list (list of path.Path): List of absolute paths of
+                user fonts.
         """
-        # display list of fonts
-        for font_file_name in font_file_name_list:
-            logger.debug("Font '%s' found to be loaded", font_file_name)
+        # get font file name
+        font_file_name = font_file_path.basename()
 
-        # load the fonts
-        for font_file_name in font_file_name_list:
-            with path(
-                "dakara_player.resources.fonts", font_file_name
-            ) as font_file_path:
-                self.load_font(Path(font_file_path))
+        # check if the font is installed at system level
+        if any(font_file_name in path for path in system_font_path_list):
+            logger.debug("Font '%s' found in system directory", font_file_name)
+            return
+
+        # check if the font is installed at user level
+        if any(font_file_name in path for path in user_font_path_list):
+            logger.debug("Font '%s' found in user directory", font_file_name)
+            return
+
+        # check if the font exists as broken link at user level
+        # in this case remove it and continue execution
+        font_file_user_path = self.FONT_DIR_USER.expanduser() / font_file_name
+        if font_file_user_path.islink():
+            logger.debug(
+                "Dead symbolic link found for font '%s' in user directory, "
+                "removing it",
+                font_file_name,
+            )
+            font_file_user_path.unlink_p()
+
+        # then, if the font is not installed, load by copying it
+        font_file_path.copy(font_file_user_path)
+
+        # register the font
+        self.fonts_loaded[font_file_name] = font_file_user_path
+
+        logger.debug(
+            "Font '%s' loaded in user directory: '%s'",
+            font_file_name,
+            font_file_user_path,
+        )
+
+    def unload(self):
+        """Remove loaded fonts."""
+        for font_file_name in self.fonts_loaded.copy():
+            self.unload_font(font_file_name)
+
+    def unload_font(self, font_file_name):
+        """Remove the provided font.
+
+        Args:
+            font_file_name (str): Name of the font to unload.
+        """
+        try:
+            font_file_path = self.fonts_loaded.pop(font_file_name)
+            font_file_path.unlink()
+            logger.debug("Font '%s' unloaded", font_file_name)
+
+        except OSError as error:
+            logger.error(
+                "Font '%s' in '%s' cannot be unloaded: %s",
+                font_file_name,
+                font_file_path,
+                error,
+            )
+
+
+class FontLoaderWindows(FontLoader):
+    """Font loader for Windows.
+
+    It uses the gdi32 library to dynamically load and unload fonts for the
+    current session.
+
+    Example of use:
+
+    >>> with FontLoaderWindows() as loader:
+    ...     loader.load()
+    ...     # do stuff while fonts are loaded
+    >>> # fonts are unloaded
+
+    Args:
+        package (str): Package checked for font files.
+
+    Attributes:
+        package (str): Package checked for font files.
+        font_loaded (dict of path.Path): List of loaded fonts. The key is the
+            font file name and the value is the path of font used at
+            installation.
+    """
+
+    GREETINGS = "Font loader for Windows selected"
+
+    def __init__(self, *args, **kwargs):
+        # call parent constructor
+        super().__init__(*args, **kwargs)
+
+        # create list of fonts
+        self.fonts_loaded = {}
+
+        # check we can use gdi32 library
+        if not hasattr(ctypes, "windll"):
+            raise FontLoaderNotAvailableError(
+                "FontLoaderWindows can only be used on Windows"
+            )
+
+    def load(self):
+        """Load the fonts."""
+        for font_file_path in self.get_font_path_iterator():
+            self.load_font(font_file_path)
 
     def load_font(self, font_file_path):
         """Load the provided font.
@@ -145,108 +268,33 @@ class FontLoaderLinux(FontLoader):
         Args:
             font_file_path (path.Path): Absolute path of the font to load.
         """
-        # get font file name
-        font_file_name = font_file_path.basename()
-
-        # check if the font is installed at system level
-        if (self.FONT_DIR_SYSTEM / font_file_name).exists():
-            logger.debug("Font '%s' found in system directory", font_file_name)
+        success = ctypes.windll.gdi32.AddFontResourceW(font_file_path)
+        if success:
+            self.fonts_loaded[font_file_path.name] = font_file_path
+            logger.debug("Font '%s' loaded", font_file_path.name)
             return
 
-        # check if the font is installed at user level
-        font_file_user_path = self.FONT_DIR_USER.expanduser() / font_file_name
-
-        if font_file_user_path.exists():
-            logger.debug("Font '%s' found in user directory", font_file_name)
-            return
-
-        # check if the font exists as broken link at user level
-        # in this case remove it and continue execution
-        if font_file_user_path.islink():
-            logger.debug(
-                "Dead symbolic link found for font '%s' in user directory, "
-                "removing it",
-                font_file_name,
-            )
-            font_file_user_path.unlink()
-
-        # then, if the font is not installed, load it
-        font_file_target_path = self.FONT_DIR_USER.expanduser() / font_file_name
-
-        font_file_path.symlink(font_file_target_path)
-
-        # register the font
-        self.fonts_loaded.append(font_file_target_path)
-
-        logger.debug(
-            "Font '%s' loaded in user directory: '%s'",
-            font_file_name,
-            font_file_target_path,
-        )
+        logger.warning("Font '%s' cannot be loaded", font_file_path.name)
 
     def unload(self):
         """Remove loaded fonts."""
-        for font_path in self.fonts_loaded.copy():
-            self.unload_font(font_path)
+        for font_file_name in self.fonts_loaded.copy():
+            self.unload_font(font_file_name)
 
-    def unload_font(self, font_path):
+    def unload_font(self, font_file_name):
         """Remove the provided font.
 
         Args:
-            font_path (str): Absolute path of the font to unload.
+            font_file_name (str): Name of the font to unload.
         """
-        try:
-            font_path.unlink()
-            self.fonts_loaded.remove(font_path)
-            logger.debug("Font '%s' unloaded", font_path)
+        font_file_path = self.fonts_loaded.pop(font_file_name)
+        success = ctypes.windll.gdi32.RemoveFontResourceW(font_file_path)
+        if success:
+            logger.debug("Font '%s' unloaded", font_file_name)
+            return
 
-        except OSError:
-            logger.error("Unable to unload '%s'", font_path)
+        logger.warning("Font '%s' cannot be unloaded", font_file_name)
 
 
-class FontLoaderWindows(FontLoader):
-    """Font loader for Windows.
-
-    It cannot do anything, since it is impossible to load fonts on Windows
-    programatically, as for now. It simply asks the user to do so.
-
-    Example of use:
-
-    >>> with FontLoaderWindows() as loader:
-    ...     loader.load() # prompts the user to install fonts manually
-    ...     # do stuff while fonts are loaded
-    >>> # fonts are not unloaded, as they were manually installed
-
-    Args:
-        output (io.BaseIO): Output stream. By default, stdout.
-    """
-
-    GREETINGS = "Font loader for Windows selected"
-
-    def __init__(self, output=None):
-        super().__init__()
-        self.output = output or sys.stdout
-
-    def load(self):
-        """Prompt the user to load the fonts.
-
-        Since there seems to be no workable way to load fonts on Windows
-        through Python, we ask the user to do it by themselve.
-        """
-        font_file_name_list = self.get_font_name_list()
-
-        self.output.write("Please install the following fonts and press Enter:\n")
-
-        # extract fonts from package if necessary and present valid paths
-        with ExitStack() as stack:
-            for font_file_name in font_file_name_list:
-                font_file_path = stack.enter_context(
-                    path("dakara_player.resources.fonts", font_file_name)
-                )
-                self.output.write("{}\n".format(font_file_path))
-
-            input()
-
-    def unload(self):
-        """Promt the user to remove the fonts."""
-        self.output.write("You can now remove the installed fonts\n")
+class FontLoaderNotAvailableError(DakaraError):
+    """Error raised when a font loader cannot be used."""