+++ tests/unit/test_font_loader.py
@@ -1,5 +1,4 @@
 import sys
-from io import StringIO
 from unittest import TestCase, skipUnless
 from unittest.mock import call, patch
 
@@ -7,6 +6,7 @@ from path import Path
 
 from dakara_player.font_loader import (
     FontLoaderLinux,
+    FontLoaderNotAvailableError,
     FontLoaderWindows,
     get_font_loader_class,
 )
@@ -36,9 +36,8 @@ class GetFontLoaderClassTestCase(TestCase):
                 get_font_loader_class()
 
 
-@skipUnless(sys.platform.startswith("linux"), "Can be tested on Linux only")
-class FontLoaderLinuxTestCase(TestCase):
-    """Test the Linux font loader."""
+class FontLoaderTestCase(TestCase):
+    """Helper to test font loader classes."""
 
     def setUp(self):
         # create directory
@@ -48,35 +47,39 @@ class FontLoaderLinuxTestCase(TestCase):
         self.font_name = "font_file.ttf"
         self.font_name_list = [self.font_name]
         self.font_path = self.directory / self.font_name
-        self.font_path_list = [self.font_path]
+        self.font_path_list = {self.font_name: self.font_path}
 
-        # save user directory
-        self.user_directory = Path("~").expanduser()
 
-        # create a font loader object
+class FontLoaderCommonTestCase(FontLoaderTestCase):
+    """Test common methods of the font loaders."""
+
+    def get_font_loader(self):
+        """Return an instance of the font loader."""
         with self.assertLogs("dakara_player.font_loader", "DEBUG"):
-            self.font_loader = FontLoaderLinux()
+            return FontLoaderLinux("package")
 
     @patch.object(FontLoaderLinux, "unload")
     def test_context_manager(self, mocked_unload):
         """Test the font loader context manager."""
-        with FontLoaderLinux():
+        with FontLoaderLinux("package"):
             pass
 
         mocked_unload.assert_called_with()
 
-    @patch.object(FontLoaderLinux, "load_from_list")
     @patch("dakara_player.font_loader.contents", autospec=True)
-    def test_load_from_resources_directory(
-        self, mocked_contents, mocked_load_from_list
-    ):
-        """Test to load fonts from the resources directory."""
+    def test_get_font_name_list(self, mocked_contents):
+        """Test to get list of font names."""
         # mock system calls
         mocked_contents.return_value = [self.font_name, "__init__.py"]
 
+        font_loader = self.get_font_loader()
+
         # call the method
         with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.load_from_resources_directory()
+            font_name_list = font_loader.get_font_name_list()
+
+        # assert the result
+        self.assertListEqual(font_name_list, self.font_name_list)
 
         # assert effect on logs
         self.assertListEqual(
@@ -88,47 +91,100 @@ class FontLoaderLinuxTestCase(TestCase):
         )
 
         # call assertions
-        mocked_contents.assert_called_once_with("dakara_player.resources.fonts")
-        mocked_load_from_list.assert_called_once_with(self.font_name_list)
+        mocked_contents.assert_called_once_with("package")
 
     @patch("dakara_player.font_loader.path")
-    @patch.object(FontLoaderLinux, "load_font")
-    def test_load_from_list(self, mocked_load_font, mocked_path):
-        """Test to load fonts from list."""
+    @patch.object(FontLoaderLinux, "get_font_name_list", autospec=True)
+    def test_get_font_path_iterator(self, mocked_get_font_name_list, mocked_path):
+        """Test to get iterator of font paths."""
+        mocked_get_font_name_list.return_value = self.font_name_list
         mocked_path.return_value.__enter__.return_value = (
             self.directory / self.font_name
         )
 
-        # call the method
-        with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.load_from_list(self.font_name_list)
+        font_loader = self.get_font_loader()
 
-        # assert effect on logs
-        self.assertListEqual(
-            logger.output,
-            ["DEBUG:dakara_player.font_loader:Font 'font_file.ttf' found to be loaded"],
-        )
+        # call the method
+        font_file_path_list = list(font_loader.get_font_path_iterator())
+
+        self.assertListEqual(font_file_path_list, [self.font_path])
+
+
+@skipUnless(sys.platform.startswith("linux"), "Can be tested on Linux only")
+class FontLoaderLinuxTestCase(FontLoaderTestCase):
+    """Test the Linux font loader."""
+
+    def setUp(self):
+        super().setUp()
+
+        # save user directory
+        self.user_directory = Path("~").expanduser()
+
+    def get_font_loader(self):
+        """Return an instance of the font loader."""
+        with self.assertLogs("dakara_player.font_loader", "DEBUG"):
+            return FontLoaderLinux("package")
+
+    @patch.object(FontLoaderLinux, "load_font", autospec=True)
+    @patch.object(FontLoaderLinux, "get_font_path_iterator", autospec=True)
+    @patch.object(Path, "walkfiles", autospec=True)
+    @patch.object(Path, "mkdir_p", autospec=True)
+    def test_load(
+        self,
+        mocked_mkdir_p,
+        mocked_walkfiles,
+        mocked_get_font_path_iterator,
+        mocked_load_font,
+    ):
+        """Test to load fonts."""
+        # prepare the mock
+        mocked_get_font_path_iterator.return_value = (p for p in [self.font_path])
+        mocked_walkfiles.side_effect = [
+            (p for p in [Path("/") / "usr" / "share" / "fonts" / "font1"]),
+            (p for p in [self.user_directory / ".fonts" / "font2"]),
+        ]
+
+        font_loader = self.get_font_loader()
+
+        # call the method
+        font_loader.load()
 
         # assert the call
-        mocked_load_font.assert_called_once_with(self.font_path)
+        mocked_mkdir_p.assert_called_once_with(self.user_directory / ".fonts")
+        mocked_walkfiles.assert_has_calls(
+            [
+                call(Path("/") / "usr" / "share" / "fonts"),
+                call(self.user_directory / ".fonts"),
+            ]
+        )
+        mocked_get_font_path_iterator.assert_called_once_with(font_loader)
+        mocked_load_font.assert_called_once_with(
+            font_loader,
+            self.font_path,
+            [Path("/") / "usr" / "share" / "fonts" / "font1"],
+            [self.user_directory / ".fonts" / "font2"],
+        )
 
     @patch.object(Path, "unlink", autospec=True)
-    @patch.object(Path, "symlink", autospec=True)
+    @patch.object(Path, "copy", autospec=True)
     @patch.object(Path, "islink", autospec=True)
-    @patch.object(Path, "exists", autospec=True)
-    def test_load_font_system(
-        self, mocked_exists, mocked_islink, mocked_symlink, mocked_unlink
-    ):
+    def test_load_font_system(self, mocked_islink, mocked_copy, mocked_unlink):
         """Test to load one font which is in system directory."""
-        # prepare the mock
-        mocked_exists.return_value = True
+        font_loader = self.get_font_loader()
 
         # pre assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # call the method
         with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.load_font(self.font_path)
+            font_loader.load_font(
+                self.font_path,
+                [Path("/") / "usr" / "share" / "fonts" / "truetype" / "font_file.ttf"],
+                [],
+            )
+
+        # post assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # assert effect on logs
         self.assertListEqual(
@@ -140,31 +196,30 @@ class FontLoaderLinuxTestCase(TestCase):
         )
 
         # assert the call
-        mocked_exists.assert_called_once_with("/usr/share/fonts/font_file.ttf")
         mocked_islink.assert_not_called()
         mocked_unlink.assert_not_called()
-        mocked_symlink.assert_not_called()
-
-        # post assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
+        mocked_copy.assert_not_called()
 
     @patch.object(Path, "unlink", autospec=True)
-    @patch.object(Path, "symlink", autospec=True)
+    @patch.object(Path, "copy", autospec=True)
     @patch.object(Path, "islink", autospec=True)
-    @patch.object(Path, "exists", autospec=True)
-    def test_load_font_user(
-        self, mocked_exists, mocked_islink, mocked_symlink, mocked_unlink
-    ):
+    def test_load_font_user(self, mocked_islink, mocked_copy, mocked_unlink):
         """Test to load one font which is in user directory."""
-        # prepare the mock
-        mocked_exists.side_effect = [False, True]
+        font_loader = self.get_font_loader()
 
         # pre assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # call the method
         with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.load_font(self.font_path)
+            font_loader.load_font(
+                self.font_path,
+                [],
+                [self.user_directory / "fonts" / "truetype" / "font_file.ttf"],
+            )
+
+        # post assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # assert effect on logs
         self.assertListEqual(
@@ -176,41 +231,34 @@ class FontLoaderLinuxTestCase(TestCase):
         )
 
         # assert the call
-        mocked_exists.assert_has_calls(
-            [
-                call("/usr/share/fonts/font_file.ttf"),
-                call(self.user_directory / ".fonts/font_file.ttf"),
-            ]
-        )
         mocked_islink.assert_not_called()
         mocked_unlink.assert_not_called()
-        mocked_symlink.assert_not_called()
-
-        # post assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
+        mocked_copy.assert_not_called()
 
     @patch.object(Path, "unlink", autospec=True)
-    @patch.object(Path, "symlink", autospec=True)
+    @patch.object(Path, "copy", autospec=True)
     @patch.object(Path, "islink", autospec=True)
-    @patch.object(Path, "exists", autospec=True)
     def test_load_font_user_link_dead_install(
         self,
-        mocked_exists,
         mocked_islink,
-        mocked_symlink,
+        mocked_copy,
         mocked_unlink,
     ):
         """Test to load one font which is in user directory as dead link."""
         # prepare the mock
-        mocked_exists.return_value = False
         mocked_islink.return_value = True
 
+        font_loader = self.get_font_loader()
+
         # pre assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # call the method
         with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.load_font(self.font_path)
+            font_loader.load_font(self.font_path, [], [])
+
+        # post assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
 
         # assert effect on logs
         font_path = self.user_directory / ".fonts/font_file.ttf"
@@ -225,37 +273,33 @@ class FontLoaderLinuxTestCase(TestCase):
         )
 
         # assert the call
-        mocked_exists.assert_has_calls(
-            [
-                call("/usr/share/fonts/font_file.ttf"),
-                call(self.user_directory / ".fonts/font_file.ttf"),
-            ]
-        )
         mocked_islink.assert_called_with(font_path)
         mocked_unlink.assert_called_with(font_path)
-        mocked_symlink.assert_called_once_with("directory/font_file.ttf", font_path)
-
-        # post assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 1)
+        mocked_copy.assert_called_once_with("directory/font_file.ttf", font_path)
 
     @patch.object(Path, "unlink", autospec=True)
-    @patch.object(Path, "symlink", autospec=True)
+    @patch.object(Path, "copy", autospec=True)
     @patch.object(Path, "islink", autospec=True)
-    @patch.object(Path, "exists", autospec=True)
-    def test_load_font_install(
-        self, mocked_exists, mocked_islink, mocked_symlink, mocked_unlink
-    ):
+    def test_load_font_install(self, mocked_islink, mocked_copy, mocked_unlink):
         """Test to load one font which is not installed."""
         # prepare the mock
-        mocked_exists.return_value = False
         mocked_islink.return_value = False
 
+        font_loader = self.get_font_loader()
+
         # pre assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # call the method
         with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.load_font(self.font_path)
+            font_loader.load_font(self.font_path, [], [])
+
+        # post assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
+        self.assertEqual(
+            font_loader.fonts_loaded[self.font_name],
+            self.user_directory / ".fonts/font_file.ttf",
+        )
 
         # assert effect on logs
         font_path = self.user_directory / ".fonts/font_file.ttf"
@@ -268,55 +312,22 @@ class FontLoaderLinuxTestCase(TestCase):
         )
 
         # assert the call
-        mocked_exists.assert_has_calls(
-            [
-                call("/usr/share/fonts/font_file.ttf"),
-                call(self.user_directory / ".fonts/font_file.ttf"),
-            ]
-        )
         mocked_islink.assert_called_with(font_path)
         mocked_unlink.assert_not_called()
-        mocked_symlink.assert_called_once_with(
+        mocked_copy.assert_called_once_with(
             "directory/font_file.ttf", self.user_directory / ".fonts/font_file.ttf"
         )
-
-        # post assertions
-        self.assertEqual(len(self.font_loader.fonts_loaded), 1)
-        self.assertEqual(
-            self.font_loader.fonts_loaded[0],
-            self.user_directory / ".fonts/font_file.ttf",
-        )
-
-    @patch.object(FontLoaderLinux, "load_from_resources_directory")
-    @patch("dakara_player.font_loader.os.mkdir", autospec=True)
-    def test_load(self, mocked_mkdir, mocked_load_from_resources_directory):
-        """Test to load fonts from main method."""
-        # call the method
-        self.font_loader.load()
-
-        # assert the call
-        mocked_mkdir.assert_called_once_with(self.user_directory / ".fonts")
-        mocked_load_from_resources_directory.assert_called_once_with()
-
-    @patch.object(FontLoaderLinux, "load_from_resources_directory")
-    @patch("dakara_player.font_loader.os.mkdir", autospec=True)
-    def test_load_directory_exists(
-        self, mocked_mkdir, mocked_load_from_resources_directory
-    ):
-        """Test to load fonts from main method."""
-        mocked_mkdir.side_effect = OSError("error message")
-
-        # call the method
-        self.font_loader.load()
 
     @patch.object(Path, "unlink", autospec=True)
     def test_unload(self, mocked_unlink):
         """Test to unload fonts."""
+        font_loader = self.get_font_loader()
+
         # set a font as loaded
-        self.font_loader.fonts_loaded = [Path("font1"), Path("font2")]
+        font_loader.fonts_loaded = {"font1": Path("font1"), "font2": Path("font2")}
 
         # call the method
-        self.font_loader.unload()
+        font_loader.unload()
 
         # assert the call
         # especially check that the unload function does not alter the list of
@@ -326,46 +337,56 @@ class FontLoaderLinuxTestCase(TestCase):
     @patch.object(Path, "unlink", autospec=True)
     def test_unload_font(self, mocked_unlink):
         """Test to unload one font."""
+        font_loader = self.get_font_loader()
+
         # set a font as loaded
-        self.font_loader.fonts_loaded = self.font_path_list
+        font_loader.fonts_loaded = self.font_path_list
 
         # pre assert there is one font loaded
-        self.assertEqual(len(self.font_loader.fonts_loaded), 1)
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
 
         # call the method
         with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
-            self.font_loader.unload_font(self.font_path)
+            font_loader.unload_font(self.font_name)
+
+        # assert there are no font loaded anymore
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # assert effect of logs
         self.assertListEqual(
             logger.output,
-            ["DEBUG:dakara_player.font_loader:Font 'directory/font_file.ttf' unloaded"],
+            ["DEBUG:dakara_player.font_loader:Font 'font_file.ttf' unloaded"],
         )
 
         # assert the call
         mocked_unlink.assert_called_once_with(self.font_path)
-
-        # assert there are no font loaded anymore
-        self.assertEqual(len(self.font_loader.fonts_loaded), 0)
 
     @patch.object(Path, "unlink", autospec=True)
     def test_unload_font_error(self, mocked_unlink):
         """Test to unload one font when unable to remove font."""
         mocked_unlink.side_effect = OSError("error message")
 
+        font_loader = self.get_font_loader()
+
         # set a font as loaded
-        self.font_loader.fonts_loaded = self.font_path_list
+        font_loader.fonts_loaded = self.font_path_list
+
+        # pre assert there is one font loaded
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
 
         # call the method
         with self.assertLogs("dakara_player.font_loader", "ERROR") as logger:
-            self.font_loader.unload_font(self.font_path)
+            font_loader.unload_font(self.font_name)
+
+        # assert there are no font loaded anymore
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
 
         # assert effect of logs
         self.assertListEqual(
             logger.output,
             [
-                "ERROR:dakara_player.font_loader:Unable to unload "
-                "'directory/font_file.ttf'"
+                "ERROR:dakara_player.font_loader:Font 'font_file.ttf' in "
+                "'directory/font_file.ttf' cannot be unloaded: error message"
             ],
         )
 
@@ -381,40 +402,165 @@ class FontLoaderWindowsTestCase(TestCase):
         self.font_name = "font_file.ttf"
         self.font_name_list = [self.font_name]
         self.font_path = self.directory / self.font_name
-        self.font_path_list = [self.font_path]
+        self.font_path_list = {self.font_name: self.font_path}
 
-    @patch("dakara_player.font_loader.path")
-    @patch("dakara_player.font_loader.input")
-    @patch("dakara_player.font_loader.contents")
-    def test_load(self, mocked_contents, mocked_input, mocked_path):
-        """Test to ask to load fonts manually."""
-        mocked_contents.return_value = [self.font_name, "__init__.py"]
-        mocked_path.return_value.__enter__.return_value = self.font_path
-        output = StringIO()
+    def get_font_loader(self):
+        """Return an instance of the font loader."""
+        with self.assertLogs("dakara_player.font_loader", "DEBUG"):
+            return FontLoaderWindows("package")
 
-        font_loader = FontLoaderWindows(output)
+    @patch("dakara_player.font_loader.ctypes", spec=[])
+    def test_init_no_windll(self, mocked_ctypes):
+        """Test to create font loader on a different OS."""
+        with self.assertRaisesRegex(
+            FontLoaderNotAvailableError, "FontLoaderWindows can only be used on Windows"
+        ):
+            self.get_font_loader()
+
+    @patch("dakara_player.font_loader.ctypes")
+    @patch.object(FontLoaderWindows, "load_font", autospec=True)
+    @patch.object(FontLoaderWindows, "get_font_path_iterator", autospec=True)
+    def test_load(self, mocked_get_font_path_iterator, mocked_load_font, mocked_ctypes):
+        """Test to load fonts."""
+        # prepare the mock
+        mocked_get_font_path_iterator.return_value = (p for p in [self.font_path])
+
+        font_loader = self.get_font_loader()
+
+        # call the method
         font_loader.load()
 
-        lines = output.getvalue().splitlines()
+        # assert the call
+        mocked_get_font_path_iterator.assert_called_once_with(font_loader)
+        mocked_load_font.assert_called_once_with(font_loader, self.font_path)
+
+    @patch("dakara_player.font_loader.ctypes")
+    def test_load_font(self, mocked_ctypes):
+        """Test to load one font."""
+        # setup mock
+        mocked_ctypes.windll.gdi32.AddFontResourceW.return_value = 1
+
+        font_loader = self.get_font_loader()
+
+        # pre assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
+
+        # call the method
+        with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
+            font_loader.load_font(self.font_path)
+
+        # post assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
+
+        # assert effect on logs
         self.assertListEqual(
-            lines,
+            logger.output,
+            ["DEBUG:dakara_player.font_loader:Font 'font_file.ttf' loaded"],
+        )
+
+        # assert the call
+        mocked_ctypes.windll.gdi32.AddFontResourceW.assert_called_once_with(
+            self.font_path
+        )
+
+    @patch("dakara_player.font_loader.ctypes")
+    def test_load_font_error(self, mocked_ctypes):
+        """Test to fail to load one font."""
+        # setup mock
+        mocked_ctypes.windll.gdi32.AddFontResourceW.return_value = 0
+
+        font_loader = self.get_font_loader()
+
+        # pre assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
+
+        # call the method
+        with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
+            font_loader.load_font(self.font_path)
+
+        # post assertions
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
+
+        # assert effect on logs
+        self.assertListEqual(
+            logger.output,
+            ["WARNING:dakara_player.font_loader:Font 'font_file.ttf' cannot be loaded"],
+        )
+
+    @patch("dakara_player.font_loader.ctypes")
+    @patch.object(FontLoaderWindows, "unload_font")
+    def test_unload(self, mocked_unload_font, mocked_ctypes):
+        """Test to unload fonts."""
+        font_loader = self.get_font_loader()
+
+        # set a font as loaded
+        font_loader.fonts_loaded = {"font1": Path("font1"), "font2": Path("font2")}
+
+        # call the method
+        font_loader.unload()
+
+        # assert the call
+        mocked_unload_font.assert_has_calls([call("font1"), call("font2")])
+
+    @patch("dakara_player.font_loader.ctypes")
+    def test_unload_font(self, mocked_ctypes):
+        """Test to unload one font."""
+        # setup mock
+        mocked_ctypes.windll.gdi32.RemoveFontResourceW.return_value = 1
+
+        font_loader = self.get_font_loader()
+
+        # set a font as loaded
+        font_loader.fonts_loaded = self.font_path_list
+
+        # pre assert there is one font loaded
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
+
+        # call the method
+        with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
+            font_loader.unload_font(self.font_name)
+
+        # assert there are no font loaded anymore
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
+
+        # assert effect of logs
+        self.assertListEqual(
+            logger.output,
+            ["DEBUG:dakara_player.font_loader:Font 'font_file.ttf' unloaded"],
+        )
+
+        # assert the call
+        mocked_ctypes.windll.gdi32.RemoveFontResourceW.assert_called_once_with(
+            self.font_path
+        )
+
+    @patch("dakara_player.font_loader.ctypes")
+    def test_unload_font_error(self, mocked_ctypes):
+        """Test to fail to unload one font."""
+        # setup mock
+        mocked_ctypes.windll.gdi32.RemoveFontResourceW.return_value = 0
+
+        font_loader = self.get_font_loader()
+
+        # set a font as loaded
+        font_loader.fonts_loaded = self.font_path_list
+
+        # pre assert there is one font loaded
+        self.assertEqual(len(font_loader.fonts_loaded), 1)
+        print(font_loader.fonts_loaded)
+
+        # call the method
+        with self.assertLogs("dakara_player.font_loader", "DEBUG") as logger:
+            font_loader.unload_font(self.font_name)
+
+        # assert there are no font loaded anymore
+        self.assertEqual(len(font_loader.fonts_loaded), 0)
+
+        # assert effect of logs
+        self.assertListEqual(
+            logger.output,
             [
-                "Please install the following fonts and press Enter:",
-                Path("directory") / "font_file.ttf",
+                "WARNING:dakara_player.font_loader:Font 'font_file.ttf' cannot "
+                "be unloaded",
             ],
         )
-
-        mocked_contents.assert_called_once_with("dakara_player.resources.fonts")
-        mocked_path.assert_called_once_with(
-            "dakara_player.resources.fonts", self.font_name
-        )
-        mocked_input.assert_called_once_with()
-
-    def test_unload(self):
-        """Test to ask to unload fonts manually."""
-        output = StringIO()
-        font_loader = FontLoaderWindows(output)
-
-        font_loader.unload()
-        lines = output.getvalue().splitlines()
-        self.assertListEqual(lines, ["You can now remove the installed fonts"])